+++ src/packages/relay/src/types.flow.js
@@ -118,30 +118,32 @@ export type RecordSourceSelectorProxy = {|
   +getPluralRootField: (fieldName: string) => ?$ReadOnlyArray<?RecordProxy>,
 |};
 
+type AnyObject = { +[key: string]: any, ... };
+
 type RecordProxy = $ReadOnly<{|
   copyFieldsFrom: (sourceRecord: RecordProxy) => void,
   getDataID: () => string,
-  getLinkedRecord: (name: string, args?: ?Object) => ?RecordProxy,
+  getLinkedRecord: (name: string, args?: ?AnyObject) => ?RecordProxy,
   getLinkedRecords: (
     name: string,
-    args?: ?Object,
+    args?: ?AnyObject,
   ) => ?$ReadOnlyArray<?RecordProxy>,
   getOrCreateLinkedRecord: (
     name: string,
     typeName: string,
-    args?: ?Object,
+    args?: ?AnyObject,
   ) => RecordProxy,
   getType: () => string,
-  getValue: (name: string, args?: ?Object) => mixed,
+  getValue: (name: string, args?: ?AnyObject) => mixed,
   setLinkedRecord: (
     record: RecordProxy,
     name: string,
-    args?: ?Object,
+    args?: ?AnyObject,
   ) => RecordProxy,
   setLinkedRecords: (
     records: $ReadOnlyArray<?RecordProxy>,
     name: string,
-    args?: ?Object,
+    args?: ?AnyObject,
   ) => RecordProxy,
-  setValue: (value: mixed, name: string, args?: ?Object) => RecordProxy,
+  setValue: (value: mixed, name: string, args?: ?AnyObject) => RecordProxy,
 |}>;